Incorporate partial passability in stats

+++ src/processing_scripts/load_habitat_access_updates.py
@@ -34,6 +34,9 @@ snapDistance = appconfig.config['CABD_DATABASE']['snap_distance']
 
 def main():
 
+    if iniSection == 'cheticamp':
+        return
+
     with appconfig.connectdb() as conn:
 
         query = f"""DROP TABLE IF EXISTS {dbTargetSchema}.{datatable};"""
